+++ src/Symfony/Component/HttpKernel/Tests/Config/FileLocatorTest.php
@@ -22,7 +22,7 @@ class FileLocatorTest extends TestCase
         $kernel
             ->expects($this->atLeastOnce())
             ->method('locateResource')
-            ->with('@BundleName/some/path', null, true)
+            ->with('@BundleName/some/path')
             ->willReturn('/bundle-name/some/path');
         $locator = new FileLocator($kernel);
         $this->assertEquals('/bundle-name/some/path', $locator->locate('@BundleName/some/path'));
@@ -32,20 +32,5 @@ class FileLocatorTest extends TestCase
             ->method('locateResource');
         $this->expectException('LogicException');
         $locator->locate('/some/path');
-    }
-
-    /**
-     * @group legacy
-     */
-    public function testLocateWithGlobalResourcePath()
-    {
-        $kernel = $this->getMockBuilder('Symfony\Component\HttpKernel\KernelInterface')->getMock();
-        $kernel
-            ->expects($this->atLeastOnce())
-            ->method('locateResource')
-            ->with('@BundleName/some/path', '/global/resource/path', false);
-
-        $locator = new FileLocator($kernel, '/global/resource/path');
-        $locator->locate('@BundleName/some/path', null, false);
     }
 }